endpoints + approvers: per-plugin files
Mirror the credentials split: each endpoint and each approver gets
its own file carrying the struct + every interface impl + the init()
that registers it. Cross-cutting helpers move to util.go.

config/plugins/endpoints/:
  endpoints.go → util.go (CredBinding helpers, validateBinding,
                          parseCredentialList, multiCredValidate,
                          singularRef, emitCredentialBinding)
                + https.go (HTTPSEndpoint + runtime + register)
                + postgres.go (PostgresEndpoint + runtime + register;
                  postgres_runtime.go and postgres_auth.go stay as
                  postgres-specific runtime impl details)
                + kubernetes.go
                + clickhouse_https.go
                + clickhouse_native.go

config/plugins/approvers/:
  approvers.go + runtime.go + llm_runtime.go
    → util.go (buildPending, decision)
      + dashboard.go (built-in DashboardApprover)
      + human.go (HumanApprover + Approve + register)
      + llm.go (LLMApprover + Approve + judge prompt + register)

Endpoint plugins now share singularRef + passthroughBuild +
multiCredValidate via util.go instead of repeating them in each
init() — endpoints/{kubernetes,clickhouse_*}.go used to inline a
copy of the same RefSpec slice that singularRef now carries.

Behaviour unchanged. tests + go build both clean.
